Partial merge of latest Cochrane fork. Little fixes and additions in audio module. Added scriptable soundtrack map size.

+++ vt/l_tr4.cpp
@@ -755,7 +755,7 @@ void TR_Level::read_tr4_level(SDL_RWops * const _src)
             this->samples_count = i;
             // Since sample data is the last part, we simply load whole last
             // block of file as single array.
-            this->samples_data_size = SDL_RWsize(src) - SDL_RWtell(src);
+            this->samples_data_size = (uint32_t) (SDL_RWsize(src) - SDL_RWtell(src));
             this->samples_data = (uint8_t*)malloc(this->samples_data_size * sizeof(uint8_t));
             for(i = 0; i < this->samples_data_size; i++)
                 this->samples_data[i] = read_bitu8(src);
